+++ test/sablono/test.cljs
@@ -4,10 +4,12 @@
             [sablono.core-test]
             [sablono.interpreter-test]
             [sablono.normalize-test]
+            [sablono.server-test]
             [sablono.util-test]))
 
 (doo-tests 'sablono.core-test
            'sablono.interpreter-test
            'sablono.normalize-test
+           'sablono.server-test
            'sablono.util-test
            'sablono.benchmark)