+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.6.0
+
+* [#100](https://github.com/r0man/sablono/issues/100) Move the
+  `render` and `render-static` functions into the `sablono.server`
+  namespace, to make the dependency on `react-dom-server` optional.
+
 ## 0.5.4
 
 ### Changes